# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel 8 OpenApi TV Channel Documentation",
        "description": "Laravel 8 OpenApi TV Channel description, Cross-platform lab 1-6",
        "contact": {
            "email": "kilubaka@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
        "/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get list of categories",
                "description": "Returns list of categories",
                "operationId": "getCategoriesList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Store new category",
                "description": "Returns created category data",
                "operationId": "storeCategory",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreCategoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Get category information",
                "description": "Returns category data",
                "operationId": "getCategoryById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Update existing category",
                "description": "Returns updated category data",
                "operationId": "updateCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCategoryRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Delete existing category",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Category id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/leaders": {
            "get": {
                "tags": [
                    "Leaders"
                ],
                "summary": "Get list of leaders",
                "description": "Returns list of leaders",
                "operationId": "getLeadersList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Leaders"
                ],
                "summary": "Store new leader",
                "description": "Returns created leader data",
                "operationId": "storeLeader",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreLeaderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/leaders/{id}": {
            "get": {
                "tags": [
                    "Leaders"
                ],
                "summary": "Get leader information",
                "description": "Returns category data",
                "operationId": "getLeaderById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Leader id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Leaders"
                ],
                "summary": "Update existing leader",
                "description": "Returns updated leader data",
                "operationId": "updateLeader",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Leader id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateLeaderRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Leaders"
                ],
                "summary": "Delete existing leader",
                "description": "Deletes a record and returns no content",
                "operationId": "deleteLeader",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Leader id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreCategoryRequest": {
                "title": "Store Category request",
                "description": "Store Category request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Name of the new category",
                        "type": "string",
                        "example": "A nice category"
                    }
                },
                "type": "object"
            },
            "StoreLeaderRequest": {
                "title": "Store Leader request",
                "description": "Store Leader request body data",
                "required": [
                    "name",
                    "date_of_birth",
                    "email"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Name of the new leader",
                        "type": "string",
                        "format": "string",
                        "example": "John Doe"
                    },
                    "date_of_birth": {
                        "title": "Date of Birth",
                        "description": "Date of Birth of the Leader",
                        "format": "date",
                        "example": "2002-01-03"
                    },
                    "email": {
                        "title": "Email",
                        "description": "Email of the Leader",
                        "type": "string",
                        "format": "string",
                        "example": "carmelo.kreigerna@okon.com"
                    }
                },
                "type": "object"
            },
...
            "UpdateCategoryRequest": {
                "title": "Update Category request",
                "description": "Update Category request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Name of the updated category",
                        "type": "string",
                        "example": "A updated category"
                    }
                },
                "type": "object"
            },
            "UpdateLeaderRequest": {
                "title": "Update Leader request",
                "description": "Update Leader request body data",
                "required": [
                    "name",
                    "date_of_birth",
                    "email"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Name of the updated leader",
                        "type": "string",
                        "format": "string",
                        "example": "John Doe"
                    },
                    "date_of_birth": {
                        "title": "Date of Birth",
                        "description": "Date of Birth of the Leader",
                        "format": "date",
                        "example": "2002-01-03"
                    },
                    "email": {
                        "title": "Email",
                        "description": "Email of the Leader",
                        "type": "string",
                        "format": "string",
                        "example": "carmelo.kreigerna@okon.com"
                    }
                },
                "type": "object"
            },
...
        }
    },
    "security": [
        []
    ]
}